+++ sources/clickup-source/src/streams/spaces.ts
@@ -1,30 +1,14 @@
-import {
-  AirbyteLogger,
-  AirbyteStreamBase,
-  StreamKey,
-  SyncMode,
-} from 'faros-airbyte-cdk';
+import {StreamKey, SyncMode} from 'faros-airbyte-cdk';
 import {Space} from 'faros-airbyte-common/clickup';
 import {Dictionary} from 'ts-essentials';
 
-import {ClickUpConfig} from '..';
-import {ClickUp} from '../clickup';
+import {StreamBase} from './common';
 
 interface StreamSlice {
   workspaceId: string;
 }
 
-export class Spaces extends AirbyteStreamBase {
-  private clickup: ClickUp;
-
-  constructor(
-    private readonly cfg: ClickUpConfig,
-    protected readonly logger: AirbyteLogger
-  ) {
-    super(logger);
-    this.clickup = ClickUp.instance(cfg, logger);
-  }
-
+export class Spaces extends StreamBase {
   getJsonSchema(): Dictionary<any, string> {
     return require('../../resources/schemas/spaces.json');
   }
