+++ sources/clickup-source/src/streams/status_histories.ts
@@ -1,34 +1,13 @@
-import {
-  AirbyteLogger,
-  AirbyteStreamBase,
-  StreamKey,
-  SyncMode,
-} from 'faros-airbyte-cdk';
+import {StreamKey, SyncMode} from 'faros-airbyte-cdk';
 import {StatusHistory} from 'faros-airbyte-common/clickup';
 import {chunk} from 'lodash';
 import {Dictionary} from 'ts-essentials';
 
-import {ClickUpConfig} from '..';
-import {ClickUp} from '../clickup';
-
-interface StreamSlice {
-  workspaceId: string;
-  listId: string;
-}
+import {ListStreamSlice, StreamWithListSlices} from './common';
 
 type StreamState = {[listId: string]: {lastUpdatedDate: string}};
 
-export class StatusHistories extends AirbyteStreamBase {
-  private clickup: ClickUp;
-
-  constructor(
-    private readonly cfg: ClickUpConfig,
-    protected readonly logger: AirbyteLogger
-  ) {
-    super(logger);
-    this.clickup = ClickUp.instance(cfg, logger);
-  }
-
+export class StatusHistories extends StreamWithListSlices {
   getJsonSchema(): Dictionary<any, string> {
     return require('../../resources/schemas/status_histories.json');
   }
@@ -41,38 +20,10 @@ export class StatusHistories extends AirbyteStreamBase {
     return ['computedProperties', 'task', 'date_updated'];
   }
 
-  async *streamSlices(): AsyncGenerator<StreamSlice> {
-    for (const workspace of await this.clickup.workspaces()) {
-      const baseSlice = {workspaceId: workspace.id};
-      for (const space of await this.clickup.spaces(
-        workspace.id,
-        this.cfg.fetch_archived
-      )) {
-        for (const list of await this.clickup.listsInSpace(
-          space.id,
-          this.cfg.fetch_archived
-        )) {
-          yield {...baseSlice, listId: list.id};
-        }
-        for (const folder of await this.clickup.folders(
-          space.id,
-          this.cfg.fetch_archived
-        )) {
-          for (const list of await this.clickup.listsInFolder(
-            folder.id,
-            this.cfg.fetch_archived
-          )) {
-            yield {...baseSlice, listId: list.id};
-          }
-        }
-      }
-    }
-  }
-
   async *readRecords(
     syncMode: SyncMode,
     cursorField?: string[],
-    streamSlice?: StreamSlice,
+    streamSlice?: ListStreamSlice,
     streamState?: StreamState
   ): AsyncGenerator<StatusHistories> {
     const listId = streamSlice.listId;
