+++ sources/clickup-source/src/streams/tasks.ts
@@ -1,33 +1,12 @@
-import {
-  AirbyteLogger,
-  AirbyteStreamBase,
-  StreamKey,
-  SyncMode,
-} from 'faros-airbyte-cdk';
+import {StreamKey, SyncMode} from 'faros-airbyte-cdk';
 import {Task} from 'faros-airbyte-common/clickup';
 import {Dictionary} from 'ts-essentials';
 
-import {ClickUpConfig} from '..';
-import {ClickUp} from '../clickup';
-
-interface StreamSlice {
-  workspaceId: string;
-  listId: string;
-}
+import {ListStreamSlice, StreamWithListSlices} from './common';
 
 type StreamState = {[listId: string]: {lastUpdatedDate: string}};
 
-export class Tasks extends AirbyteStreamBase {
-  private clickup: ClickUp;
-
-  constructor(
-    private readonly cfg: ClickUpConfig,
-    protected readonly logger: AirbyteLogger
-  ) {
-    super(logger);
-    this.clickup = ClickUp.instance(cfg, logger);
-  }
-
+export class Tasks extends StreamWithListSlices {
   getJsonSchema(): Dictionary<any, string> {
     return require('../../resources/schemas/tasks.json');
   }
@@ -40,38 +19,10 @@ export class Tasks extends AirbyteStreamBase {
     return 'date_updated';
   }
 
-  async *streamSlices(): AsyncGenerator<StreamSlice> {
-    for (const workspace of await this.clickup.workspaces()) {
-      const baseSlice = {workspaceId: workspace.id};
-      for (const space of await this.clickup.spaces(
-        workspace.id,
-        this.cfg.fetch_archived
-      )) {
-        for (const list of await this.clickup.listsInSpace(
-          space.id,
-          this.cfg.fetch_archived
-        )) {
-          yield {...baseSlice, listId: list.id};
-        }
-        for (const folder of await this.clickup.folders(
-          space.id,
-          this.cfg.fetch_archived
-        )) {
-          for (const list of await this.clickup.listsInFolder(
-            folder.id,
-            this.cfg.fetch_archived
-          )) {
-            yield {...baseSlice, listId: list.id};
-          }
-        }
-      }
-    }
-  }
-
   async *readRecords(
     syncMode: SyncMode,
     cursorField?: string[],
-    streamSlice?: StreamSlice,
+    streamSlice?: ListStreamSlice,
     streamState?: StreamState
   ): AsyncGenerator<Task> {
     const listId = streamSlice.listId;
